Rename ts_type module
