+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/ts_fn_or_constructor_type/ts_fn_type/mod.rs
