+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/ts_keyword_type/errors.rs
